requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg<0.29.0
pgvector==0.3.6
alembic==1.14.0

# Settings & validation
pydantic-settings==2.6.1

# Auth
python-jose[cryptography]==3.3.0
google-auth==2.37.0
requests==2.32.3

# HTTP client
httpx==0.28.1

# AI
openai==1.59.3
